refactor: move button to OnboardingButton component
