+++ app/components/OnboardingButton/styles.ts
@@ -1,0 +1,31 @@
+import { createStyleSheet } from 'react-native-unistyles';
+
+export const stylesheet = createStyleSheet({
+  button: {
+    marginTop: 16,
+    backgroundColor: '#EF7E06',
+    borderRadius: 100,
+    width: '100%',
+    height: 64,
+    alignItems: 'center',
+    justifyContent: 'center',
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 3,
+  },
+  buttonText: {
+    color: 'white',
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  wrapper: {
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingHorizontal: 16,
+  },
+});
