+++ app/components/index.ts
@@ -2,3 +2,4 @@ export * from './FavoritesHeader/FavoritesHeader';
 export * from './BottomSheet/BottomSheet';
 export * from './OnboardingMask/OnboardingMask';
 export * from './OnboardingPagination/OnboardingPagination';
+export * from './OnboardingButton/OnboardingButton';
