+++ app/screens/OnboardingScreen/OnboardingScreen.tsx
@@ -1,60 +1,25 @@
 import { useState } from 'react';
-import {
-  SafeAreaView,
-  View,
-  Text,
-  TouchableOpacity,
-  Image,
-} from 'react-native';
+import { SafeAreaView, View, Text, Image } from 'react-native';
 import { useStyles } from 'react-native-unistyles';
+import { useSharedValue } from 'react-native-reanimated';
 import { useTranslation } from 'react-i18next';
-import AntDesign from 'react-native-vector-icons/AntDesign';
-import Animated, {
-  useAnimatedStyle,
-  withSpring,
-  withTiming,
-  useSharedValue,
-} from 'react-native-reanimated';
-import { trigger } from 'react-native-haptic-feedback';
-import { storage } from '@config/storage';
-import { stylesheet } from './styles';
 import {
-  HAPTIC_FEEDBACK_OPTIONS,
-  SCREEN_DIMENSIONS,
+  OnboardingMask,
+  OnboardingPagination,
+  OnboardingButton,
+} from '@components';
+import {
   ONBOARDING_BACKGROUND_COLORS,
   ONBOARDING_IMAGES,
   ONBOARDING_TITLES,
-  ONBOARDING_TOTAL_STEPS,
 } from '@config/constants';
-import { OnboardingMask, OnboardingPagination } from '@components';
+import { stylesheet } from './styles';
 
 export const OnboardingScreen = () => {
   const [step, setStep] = useState<number>(1);
   const { styles } = useStyles(stylesheet);
-  const { t } = useTranslation();
   const mask = useSharedValue(0);
-
-  const handleSetOnboarding = async () => {
-    trigger('impactLight', HAPTIC_FEEDBACK_OPTIONS);
-
-    if (step < ONBOARDING_TOTAL_STEPS) {
-      setStep(prevState => prevState + 1);
-
-      if (step === 2) mask.value = 0;
-      mask.value = withTiming(SCREEN_DIMENSIONS.height, { duration: 1000 });
-    } else {
-      storage.set('isOnboardingComplete', true);
-    }
-  };
-
-  const buttonAnimationStyle = useAnimatedStyle(() => {
-    return {
-      width:
-        step === ONBOARDING_TOTAL_STEPS ? withSpring(278) : withSpring(128),
-      height:
-        step === ONBOARDING_TOTAL_STEPS ? withSpring(64) : withSpring(128),
-    };
-  });
+  const { t } = useTranslation();
 
   return (
     <SafeAreaView
@@ -75,18 +40,8 @@ export const OnboardingScreen = () => {
         <Text style={styles.title}>{t(ONBOARDING_TITLES[step - 1])}</Text>
       </View>
 
-      <View style={styles.wrapper}>
-        <TouchableOpacity activeOpacity={0.7} onPress={handleSetOnboarding}>
-          <Animated.View style={[styles.button, buttonAnimationStyle]}>
-            {step < ONBOARDING_TOTAL_STEPS ? (
-              <AntDesign name="arrowright" size={42} color="white" />
-            ) : (
-              <Text style={styles.buttonText}>{t('onboardingButton')}</Text>
-            )}
-          </Animated.View>
-        </TouchableOpacity>
-      </View>
-
+      <OnboardingButton step={step} setStep={setStep} mask={mask} />
+      
       <OnboardingPagination step={step} />
     </SafeAreaView>
   );
