+++ app/screens/OnboardingScreen/styles.ts
@@ -28,28 +28,6 @@ export const stylesheet = createStyleSheet({
     fontWeight: 'bold',
     textAlign: 'left',
   },
-  button: {
-    marginTop: 16,
-    backgroundColor: '#EF7E06',
-    borderRadius: 100,
-    width: '100%',
-    height: 64,
-    alignItems: 'center',
-    justifyContent: 'center',
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 3,
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
   wrapper: {
     alignItems: 'center',
     justifyContent: 'space-between',
